initial support bidirectional, server-initiated channels
Channels need not be exclusively client-created, in fact few (if any?)
channels specify directionality, and refer only to a sender and receiver.

This does an initial, somewhat minimal pass at making servers capable of
initiating channels. This is rather rough and I think greater
abstraction is warranted here; I pulled out relevant parsing code into
`parsing.rs`, but there is logical duplication between the client and
server which I don't have time to disentagle. This also only implements
_some_ channel functionality on the server-side, it's not as complete as
the client.

The significant breaking change here is that `server::run_stream` now
returns (almost) immediatel with a SessionHandle, similar to the client;
this is used to open channels.

Ultimately this is a big set of changes I would understand if there's
not great desire to merge this upstream :)