add flags for openssl-only crypto
Fixes #50

This introduces an on-by-default `rs-crypto` flag, which enables the
existing Rust-based crypto libraries (including aes and ED25519).
However, these implementations can be removed by disabling the flag.
If it's disabled, then openssl (when turned on) will stand in for them,
in a less performant way.

Note that while OpenSSL 3.x does have some ED25519 support, I have not
done the work to make that compatible as well--partly because ED25519
is not yet an approved algorithm for my company to use, and partly to
retain compatibility with OpenSSL 1.x