impl Drop for server examples (#376)
In #373 I raised my confusion about the lack of a disconnection
mechanism in the source code and the examples, but it was pointed out
that `Drop` is the intended way of cleaning up resources.

This PR adds a Drop implementation for three of the server examples, so
that users that refer to these won't be confused about how to implement
their own cleanup.