Remove unused dependencies (#488)
I ran cargo-udeps on this workspace (on a Linux machine) and looked at
the results.

- pageant is an empty crate on non-Windows, so all of its dependencies
are Windows-only
- russh uses tokio-stream only in a doc-test. For this, dev-dependencies
suffices. I do not see why this is "not wasm32", but it already was this
way before, so I leave it there.
- russh-util does not seem to use futures-executor or static_assertions
and apparently never did (??). These were added in commit d03df31af. In
this commit, I am dropping them.
- russh-util uses chrono only on wasm

Signed-off-by: Uli Schlachter <psychon@znc.in>