fix: allow `Stream` to connect to multiple addresses
Previously, when using `config.stream()`, it would resolve the host,
port pair and pick the first address returned. But Tokio has built-in
support to resolve multiple addresses and connect to each one until
one succeeds. So this PR lifts the previous restriction.

The main consquence is that the `Error::NotResolvable` is gone since it
will be reported as an `Io` error by Tokio instead.