Fix channels being closed prematurely
Closes #553.

Previously, the server session would immediately set the EOF or
close message on the encrypted channel's session, which could
result in a race condition between it and the last data message,
sporadically causing the latter to be dropped.

This change makes the behavior match that of the client session,
setting a flag for flushing in case there's pending data.